+++ CHANGELOG.md
@@ -8,6 +8,8 @@
     * Add fractal-server version to logs (\#2228).
 * Database:
     * Remove `run_migrations_offline` from `env.py` and make `run_migrations_online` sync (\#2239).
+* Runner
+    * Sudo/SLURM executor checks the fractal-server version using `FRACTAL_SLURM_WORKER_PYTHON` config variable, if set (\#2240).
 
 # 2.12.0